Changed Amount from int to decimal

+++ ProjectLedg.Server/Data/Models/InvoiceItems.cs
@@ -1,4 +1,4 @@
-﻿using System.ComponentModel.DataAnnotations;
+﻿    using System.ComponentModel.DataAnnotations;
 
 namespace ProjectLedg.Server.Data.Models
 {
@@ -9,7 +9,7 @@ namespace ProjectLedg.Server.Data.Models
         public string Description { get; set; }
         public int Quantity { get; set; }
         public decimal UnitPrice { get; set; }
-        public int Amount { get; set; }
+        public decimal Amount { get; set; }
 
         public int InvoiceId { get; set; } // FK to the Invoice model
         public Invoice Invoice { get; set; } // Nav property to Invoice
